+++ src/Attribute/ChainModifier/NoOp.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Nandan108\DtoToolkit\Attribute\ChainModifier;
 
-use Nandan108\DtoToolkit\Traits\UsesParamResolver;
-
 /**
  * This is suggar for Wrap(0): a simple no-op that doesn't modify the chain.
  *
@@ -14,8 +12,6 @@ use Nandan108\DtoToolkit\Traits\UsesParamResolver;
 #[\Attribute(\Attribute::TARGET_PROPERTY | \Attribute::IS_REPEATABLE)]
 class NoOp extends Wrap
 {
-    use UsesParamResolver;
-
     public function __construct()
     {
         parent::__construct(count: 0);