+++ src/CastTo/DateTimeFromLocalized.php
@@ -35,11 +35,11 @@ final class DateTimeFromLocalized extends CastBase implements CasterInterface, B
      * @param mixed $timezone
      */
     public function __construct(
-        ?string $locale = null,
+        string | callable | null $locale = null,
         int $dateStyle = \IntlDateFormatter::SHORT,
         int $timeStyle = \IntlDateFormatter::SHORT,
         ?string $pattern = null,
-        ?string $timezone = null,
+        string | callable | null $timezone = null,
     ) {
         $this->ensureExtensionLoaded('intl');